requirements.txt
PyYAML==6.0
pypresence==4.2.1
console==0.9907
discord==1.7.3
discord.py==1.7.3